+++ CHANGES.rst
@@ -7,7 +7,8 @@ ChangeLog
 
 .. note:: This version is not yet released and is under active development.
 
-* No changes yet.
+* Reject mails without ``Message-ID`` headers if ``--message-id`` option is
+  active.
 
 
 `2.0.0 (2016-11-13) <https://github.com/kdeldycke/maildir-deduplicate/compare/v1.3.0...v2.0.0>`_